+++ graphics/r_shaders.h
@@ -3,6 +3,7 @@
 #pragma once
 
 #include "r_types.h"
+#include "fresa_config.h"
 
 namespace fresa::graphics
 {
@@ -11,20 +12,55 @@ namespace fresa::graphics
     // ···············
 
     //: shader types
-    enum struct ShaderStage {
+    struct ShaderStageData {
+        str_view extension;
+        VkShaderStageFlagBits stage;
+    };
+
+    enum struct ShaderStages {
         VERTEX,
         FRAGMENT,
         COMPUTE
     };
 
-    //: shader extensions
-    constexpr auto shader_extensions = std::to_array<str_view>({ ".vert", ".frag", ".comp" });
+    constexpr inline auto shader_stages = std::to_array<ShaderStageData>({
+        {"vert", VK_SHADER_STAGE_VERTEX_BIT},
+        {"frag", VK_SHADER_STAGE_FRAGMENT_BIT},
+        {"comp", VK_SHADER_STAGE_COMPUTE_BIT}
+    });
 
     // ·············
     // · DATATYPES ·
     // ·············
 
-    //: ...
+    //: shader resource
+    //      a binding for a specific descriptor resource item
+    /*struct ShaderResource {
+        ui16 id;
+        str name;
+        VkDescriptorType type;
+    };*/
+    
+    //: descriptor set
+    //      holds the descriptors and bindings for each glsl set, used for sending data to the shader
+    /*struct DescriptorSet {
+        ui32 set;
+        VkDescriptorSetLayout layout;
+        VkDescriptorPool* pool;
+        std::array<VkDescriptorSet, engine_config.vk_frames_in_flight()> descriptors;
+        std::vector<VkDescriptorSetLayoutBinding> bindings;
+        std::vector<ShaderResource> resources;
+    };*/
+
+    //: shader pass
+    //      combination of multiple shader stages to create a pipeline
+    //      holds the list of modules, descriptor sets and the built pipeline
+    /*struct ShaderPass {
+        std::vector<VkShaderModule> stages;
+        std::vector<DescriptorSet> descriptors;
+        VkPipelineLayout pipeline_layout;
+        VkPipeline pipeline;
+    };*/
 
     // ···········
     // · SYSTEMS ·
@@ -33,6 +69,12 @@ namespace fresa::graphics
     namespace shader
     {
         //: read spirv code
-        std::vector<ui32> readSPIRV(str name, ShaderStage type);
+        std::vector<ui32> readSPIRV(str name, ShaderStages stage);
+
+        //: create shader module from spirv code
+        VkShaderModule createModule(str name, ShaderStages stage);
+
+        //: create a shader pass from one or more shader modules, also does reflection on descriptor layouts
+        //- ShaderPass createPass(str name);
     }
 }